Fixed the conversion to JSON
We need to fix the exception when there is no parameters -f
