+++ cli/src/main/java/com/michelin/ns4kafka/cli/models/ObjectMeta.java
@@ -1,0 +1,34 @@
+package com.michelin.ns4kafka.cli.models;
+
+import java.util.Map;
+
+public class ObjectMeta {
+	private String name;
+	private String namespace;
+	private String cluster;
+	private Map<String,String> labels;
+	public String getName() {
+		return name;
+	}
+	public void setName(String name) {
+		this.name = name;
+	}
+	public String getNamespace() {
+		return namespace;
+	}
+	public void setNamespace(String namespace) {
+		this.namespace = namespace;
+	}
+	public String getCluster() {
+		return cluster;
+	}
+	public void setCluster(String cluster) {
+		this.cluster = cluster;
+	}
+	public Map<String, String> getLabels() {
+		return labels;
+	}
+	public void setLabels(Map<String, String> labels) {
+		this.labels = labels;
+	}
+}
