+++ cli/src/main/java/com/michelin/ns4kafka/cli/models/Resource.java
@@ -1,13 +1,17 @@
 package com.michelin.ns4kafka.cli.models;
 
-import java.util.Map;
-
 public class Resource {
 
-    private final String apiVersion = "v1";
+    private String apiVersion;
     private String kind;
     private ObjectMeta metadata;
     private Object spec;
+	public String getApiVersion() {
+		return apiVersion;
+	}
+	public void setApiVersion(String apiVersion) {
+		this.apiVersion = apiVersion;
+	}
 	public String getKind() {
 		return kind;
 	}
@@ -27,34 +31,4 @@ public class Resource {
 		this.spec = spec;
 	}
 
-	public class ObjectMeta {
-		private String name;
-		private String namespace;
-		private String cluster;
-		private Map<String,String> labels;
-		public String getName() {
-			return name;
-		}
-		public void setName(String name) {
-			this.name = name;
-		}
-		public String getNamespace() {
-			return namespace;
-		}
-		public void setNamespace(String namespace) {
-			this.namespace = namespace;
-		}
-		public String getCluster() {
-			return cluster;
-		}
-		public void setCluster(String cluster) {
-			this.cluster = cluster;
-		}
-		public Map<String, String> getLabels() {
-			return labels;
-		}
-		public void setLabels(Map<String, String> labels) {
-			this.labels = labels;
-		}
-	}
 }
